requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Trading/APIs
python-binance>=1.0.19
ccxt>=4.0.0
requests>=2.31.0
aiohttp>=3.9.0
websockets>=11.0.0

# Database
redis>=5.0.0
sqlalchemy>=2.0.0

# Data handling
orjson>=3.9.0

# Utilities
pydantic>=2.0.0
tenacity>=8.0.0
schedule>=1.2.0
pytz>=2024.1

# Monitoring & Logging
prometheus-client>=0.19.0

# Development/Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.0
